requirements.txt
tensorflow==2.21.0
tensorboard
keras==3.14.0
scikit-learn==1.8.0
pandas==3.0.2
numpy==2.4.4
matplotlib==3.10.8
seaborn==0.13.2
fastapi==0.135.3
uvicorn==0.43.0
python-multipart
streamlit==1.56.0
requests
joblib==1.5.3
pyyaml==6.0.3
keras-tuner>=1.4
shap>=0.51
pytest>=7.0
httpx>=0.24
